fix bug

+++ code/src/main/java/org/example/code/DTO/SanPhamBienTheDTO.java
@@ -2,9 +2,9 @@ package org.example.code.DTO;
 
 public class SanPhamBienTheDTO {
     public Integer idSanPham;
-    public Long soLuongBienThe;
+    public Integer soLuongBienThe;
 
-    public SanPhamBienTheDTO(Integer idSanPham, Long soLuongBienThe) {
+    public SanPhamBienTheDTO(Integer idSanPham, Integer soLuongBienThe) {
         this.idSanPham = idSanPham;
         this.soLuongBienThe = soLuongBienThe;
     }
